feat: added Privy React helper
